[ty] Recognize exhaustive matches over finite tuples
