+++ crates/ty_python_semantic/resources/mdtest/exhaustiveness_checking.md
@@ -92,6 +92,61 @@ def exhaustiveness_using_containment_checks():
         assert_never(norm_str)
 ```
 
+## Checks on fixed-length tuples
+
+Finite tuple elements are expanded into their Cartesian product when checking whether a `match`
+statement is exhaustive. The expansion is used only for reachability; it does not affect the
+inferred type of the subject or pattern captures.
+
+```py
+def constructed_bool_tuple(b_1: bool, b_2: bool) -> int:
+    match (b_1, b_2):
+        case (True, True):
+            return 0
+        case (False, _):
+            return 1
+        case (_, False):
+            return 2
+
+def direct_bool_tuple(pair: tuple[bool, bool]) -> int:
+    match pair:
+        case (True, True):
+            return 0
+        case (False, _):
+            return 1
+        case (_, False):
+            return 2
+
+def non_exhaustive_bool_tuple(pair: tuple[bool, bool]) -> int:  # error: [invalid-return-type]
+    match pair:
+        case (True, True):
+            return 0
+        case (False, _):
+            return 1
+
+def guarded_bool_tuple(pair: tuple[bool, bool], flag: bool) -> int:  # error: [invalid-return-type]
+    match pair:
+        case (True, True) if flag:
+            return 0
+        case (False, _):
+            return 1
+        case (_, False):
+            return 2
+
+# Expanding this tuple would produce 128 alternatives, exceeding the limit of 64. The checker
+# falls back to its conservative behavior instead of performing an exponential expansion.
+def tuple_exceeding_expansion_limit(
+    value: tuple[bool, bool, bool, bool, bool, bool, bool],
+) -> int:  # error: [invalid-return-type]
+    match value:
+        case (True, True, _, _, _, _, _):
+            return 0
+        case (False, _, _, _, _, _, _):
+            return 1
+        case (_, False, _, _, _, _, _):
+            return 2
+```
+
 ## Checks on enum literals
 
 ```py
